requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
future==0.17.1
idna==2.8
pycparser==2.19
python-telegram-bot==11.1.0
requests==2.21.0
six==1.12.0
soupsieve==1.8
spotify-token==0.1.0
urllib3==1.24.1
